More fixes (still 127 to go...)
