+++ packages/dev/core/src/Engines/engine.ts
@@ -33,16 +33,6 @@ import type { PostProcess } from "../PostProcesses/postProcess";
 import { AbstractEngine } from "./abstractEngine";
 
 /**
- * Defines the interface used by objects containing a viewport (like a camera)
- */
-interface IViewportOwnerLike {
-    /**
-     * Gets or sets the viewport
-     */
-    viewport: IViewportLike;
-}
-
-/**
  * The engine class is responsible for interfacing with all lower-level APIs such as WebGL and Audio
  */
 export class Engine extends AbstractEngine {
@@ -393,36 +383,6 @@ export class Engine extends AbstractEngine {
     public disableContextMenu: boolean = true;
 
     /**
-     * Gets the list of created scenes
-     */
-    public scenes: Scene[] = [];
-
-    /** @internal */
-    public _virtualScenes = new Array<Scene>();
-
-    /**
-     * Event raised when a new scene is created
-     */
-    public onNewSceneAddedObservable = new Observable<Scene>();
-
-    // Observables
-
-    /**
-     * Observable event triggered each time the rendering canvas is resized
-     */
-    public onResizeObservable = new Observable<Engine>();
-
-    /**
-     * Observable event triggered each time the canvas gains focus
-     */
-    public onCanvasFocusObservable = new Observable<Engine>();
-
-    /**
-     * Observable event triggered each time the canvas receives pointerout event
-     */
-    public onCanvasPointerOutObservable = new Observable<PointerEvent>();
-
-    /**
      * If set, will be used to request the next animation frame for the render loop
      */
     public customAnimationFrameRequester: Nullable<ICustomAnimationFrameRequester> = null;
@@ -452,24 +412,17 @@ export class Engine extends AbstractEngine {
     public static OfflineProviderFactory: (urlToScene: string, callbackManifestChecked: (checked: boolean) => any, disableManifestCheck: boolean) => IOfflineProvider;
 
     private _loadingScreen: ILoadingScreen;
-    private _pointerLockRequested: boolean;
     private _rescalePostProcess: Nullable<PostProcess>;
-
-    // Deterministic lockstepMaxSteps
-    protected _deterministicLockstep: boolean = false;
-    protected _lockstepMaxSteps: number = 4;
-    protected _timeStep: number = 1 / 60;
 
     protected get _supportsHardwareTextureRescaling() {
         return !!Engine._RescalePostProcessFactory;
     }
 
-    // FPS
-    private _fps = 60;
-    private _deltaTime = 0;
-
-    /** @internal */
-    public _drawCalls = new PerfCounter();
+    private _measureFps(): void {
+        this._performanceMonitor.sampleFrame();
+        this._fps = this._performanceMonitor.averageFPS;
+        this._deltaTime = this._performanceMonitor.instantaneousFrameTime || 0;
+    }
 
     /**
      * Turn this value on if you want to pause FPS computation when in background
@@ -628,56 +581,6 @@ export class Engine extends AbstractEngine {
         this._timeStep = this._creationOptions.timeStep || 1 / 60;
     }
 
-    /**
-     * Gets current aspect ratio
-     * @param viewportOwner defines the camera to use to get the aspect ratio
-     * @param useScreen defines if screen size must be used (or the current render target if any)
-     * @returns a number defining the aspect ratio
-     */
-    public getAspectRatio(viewportOwner: IViewportOwnerLike, useScreen = false): number {
-        const viewport = viewportOwner.viewport;
-        return (this.getRenderWidth(useScreen) * viewport.width) / (this.getRenderHeight(useScreen) * viewport.height);
-    }
-
-    /**
-     * Gets current screen aspect ratio
-     * @returns a number defining the aspect ratio
-     */
-    public getScreenAspectRatio(): number {
-        return this.getRenderWidth(true) / this.getRenderHeight(true);
-    }
-
-    /**
-     * Gets a boolean indicating that the engine is running in deterministic lock step mode
-     * @see https://doc.babylonjs.com/features/featuresDeepDive/animation/advanced_animations#deterministic-lockstep
-     * @returns true if engine is in deterministic lock step mode
-     */
-    public isDeterministicLockStep(): boolean {
-        return this._deterministicLockstep;
-    }
-
-    /**
-     * Gets the max steps when engine is running in deterministic lock step
-     * @see https://doc.babylonjs.com/features/featuresDeepDive/animation/advanced_animations#deterministic-lockstep
-     * @returns the max steps
-     */
-    public getLockstepMaxSteps(): number {
-        return this._lockstepMaxSteps;
-    }
-
-    /**
-     * Returns the time in ms between steps when using deterministic lock step.
-     * @returns time step in (ms)
-     */
-    public getTimeStep(): number {
-        return this._timeStep * 1000;
-    }
-
-    /**
-     * Force the mipmap generation for the given render target texture
-     * @param texture defines the render target texture to use
-     * @param unbind defines whether or not to unbind the texture after generation. Defaults to true.
-     */
     public generateMipMapsForCubemap(texture: InternalTexture, unbind = true) {
         if (texture.generateMipMaps) {
             const gl = this._gl;
@@ -690,134 +593,6 @@ export class Engine extends AbstractEngine {
     }
 
     /** States */
-
-    /**
-     * Gets a boolean indicating if stencil buffer is enabled
-     * @returns the current stencil buffer state
-     */
-    public getStencilBuffer(): boolean {
-        return this._stencilState.stencilTest;
-    }
-
-    /**
-     * Enable or disable the stencil buffer
-     * @param enable defines if the stencil buffer must be enabled or disabled
-     */
-    public setStencilBuffer(enable: boolean): void {
-        this._stencilState.stencilTest = enable;
-    }
-
-    /**
-     * Gets the current stencil mask
-     * @returns a number defining the new stencil mask to use
-     */
-    public getStencilMask(): number {
-        return this._stencilState.stencilMask;
-    }
-
-    /**
-     * Sets the current stencil mask
-     * @param mask defines the new stencil mask to use
-     */
-    public setStencilMask(mask: number): void {
-        this._stencilState.stencilMask = mask;
-    }
-
-    /**
-     * Gets the current stencil function
-     * @returns a number defining the stencil function to use
-     */
-    public getStencilFunction(): number {
-        return this._stencilState.stencilFunc;
-    }
-
-    /**
-     * Gets the current stencil reference value
-     * @returns a number defining the stencil reference value to use
-     */
-    public getStencilFunctionReference(): number {
-        return this._stencilState.stencilFuncRef;
-    }
-
-    /**
-     * Gets the current stencil mask
-     * @returns a number defining the stencil mask to use
-     */
-    public getStencilFunctionMask(): number {
-        return this._stencilState.stencilFuncMask;
-    }
-
-    /**
-     * Sets the current stencil function
-     * @param stencilFunc defines the new stencil function to use
-     */
-    public setStencilFunction(stencilFunc: number) {
-        this._stencilState.stencilFunc = stencilFunc;
-    }
-
-    /**
-     * Sets the current stencil reference
-     * @param reference defines the new stencil reference to use
-     */
-    public setStencilFunctionReference(reference: number) {
-        this._stencilState.stencilFuncRef = reference;
-    }
-
-    /**
-     * Sets the current stencil mask
-     * @param mask defines the new stencil mask to use
-     */
-    public setStencilFunctionMask(mask: number) {
-        this._stencilState.stencilFuncMask = mask;
-    }
-
-    /**
-     * Gets the current stencil operation when stencil fails
-     * @returns a number defining stencil operation to use when stencil fails
-     */
-    public getStencilOperationFail(): number {
-        return this._stencilState.stencilOpStencilFail;
-    }
-
-    /**
-     * Gets the current stencil operation when depth fails
-     * @returns a number defining stencil operation to use when depth fails
-     */
-    public getStencilOperationDepthFail(): number {
-        return this._stencilState.stencilOpDepthFail;
-    }
-
-    /**
-     * Gets the current stencil operation when stencil passes
-     * @returns a number defining stencil operation to use when stencil passes
-     */
-    public getStencilOperationPass(): number {
-        return this._stencilState.stencilOpStencilDepthPass;
-    }
-
-    /**
-     * Sets the stencil operation to use when stencil fails
-     * @param operation defines the stencil operation to use when stencil fails
-     */
-    public setStencilOperationFail(operation: number): void {
-        this._stencilState.stencilOpStencilFail = operation;
-    }
-
-    /**
-     * Sets the stencil operation to use when depth fails
-     * @param operation defines the stencil operation to use when depth fails
-     */
-    public setStencilOperationDepthFail(operation: number): void {
-        this._stencilState.stencilOpDepthFail = operation;
-    }
-
-    /**
-     * Sets the stencil operation to use when stencil passes
-     * @param operation defines the stencil operation to use when stencil passes
-     */
-    public setStencilOperationPass(operation: number): void {
-        this._stencilState.stencilOpStencilDepthPass = operation;
-    }
 
     /**
      * Sets a boolean indicating if the dithering state is enabled or disabled
@@ -1150,40 +925,6 @@ export class Engine extends AbstractEngine {
     }
 
     /**
-     * Toggle full screen mode
-     * @param requestPointerLock defines if a pointer lock should be requested from the user
-     */
-    public switchFullscreen(requestPointerLock: boolean): void {
-        if (this.isFullscreen) {
-            this.exitFullscreen();
-        } else {
-            this.enterFullscreen(requestPointerLock);
-        }
-    }
-
-    /**
-     * Enters full screen mode
-     * @param requestPointerLock defines if a pointer lock should be requested from the user
-     */
-    public enterFullscreen(requestPointerLock: boolean): void {
-        if (!this.isFullscreen) {
-            this._pointerLockRequested = requestPointerLock;
-            if (this._renderingCanvas) {
-                Engine._RequestFullscreen(this._renderingCanvas);
-            }
-        }
-    }
-
-    /**
-     * Exits full screen mode
-     */
-    public exitFullscreen(): void {
-        if (this.isFullscreen) {
-            Engine._ExitFullscreen();
-        }
-    }
-
-    /**
      * Enters Pointerlock mode
      */
     public enterPointerlock(): void {
@@ -1356,59 +1097,6 @@ export class Engine extends AbstractEngine {
         });
     }
 
-    // eslint-disable-next-line @typescript-eslint/naming-convention
-    protected static _RenderPassIdCounter = 0;
-
-    private _renderPassNames: string[] = ["main"];
-    /**
-     * Gets the names of the render passes that are currently created
-     * @returns list of the render pass names
-     */
-    public getRenderPassNames(): string[] {
-        return this._renderPassNames;
-    }
-
-    /**
-     * Gets the name of the current render pass
-     * @returns name of the current render pass
-     */
-    public getCurrentRenderPassName(): string {
-        return this._renderPassNames[this.currentRenderPassId];
-    }
-
-    /**
-     * Creates a render pass id
-     * @param name Name of the render pass (for debug purpose only)
-     * @returns the id of the new render pass
-     */
-    public createRenderPassId(name?: string) {
-        // Note: render pass id == 0 is always for the main render pass
-        const id = ++Engine._RenderPassIdCounter;
-        this._renderPassNames[id] = name ?? "NONAME";
-        return id;
-    }
-
-    /**
-     * Releases a render pass id
-     * @param id id of the render pass to release
-     */
-    public releaseRenderPassId(id: number): void {
-        this._renderPassNames[id] = undefined as any;
-
-        for (let s = 0; s < this.scenes.length; ++s) {
-            const scene = this.scenes[s];
-            for (let m = 0; m < scene.meshes.length; ++m) {
-                const mesh = scene.meshes[m];
-                if (mesh.subMeshes) {
-                    for (let b = 0; b < mesh.subMeshes.length; ++b) {
-                        const subMesh = mesh.subMeshes[b];
-                        subMesh._removeDrawWrapper(id);
-                    }
-                }
-            }
-        }
-    }
-
     /**
      * @internal
      * Rescales a texture
@@ -1467,30 +1155,6 @@ export class Engine extends AbstractEngine {
                 }
             });
         }
-    }
-
-    // FPS
-
-    /**
-     * Gets the current framerate
-     * @returns a number representing the framerate
-     */
-    public getFps(): number {
-        return this._fps;
-    }
-
-    /**
-     * Gets the time spent between current and previous frame
-     * @returns a number representing the delta time in ms
-     */
-    public getDeltaTime(): number {
-        return this._deltaTime;
-    }
-
-    private _measureFps(): void {
-        this._performanceMonitor.sampleFrame();
-        this._fps = this._performanceMonitor.averageFPS;
-        this._deltaTime = this._performanceMonitor.instantaneousFrameTime || 0;
     }
 
     /**
@@ -1865,31 +1529,6 @@ export class Engine extends AbstractEngine {
     static _ExitPointerlock(): void {
         if (document.exitPointerLock) {
             document.exitPointerLock();
-        }
-    }
-
-    /**
-     * Ask the browser to promote the current element to fullscreen rendering mode
-     * @param element defines the DOM element to promote
-     */
-    static _RequestFullscreen(element: HTMLElement): void {
-        const requestFunction = element.requestFullscreen || (<any>element).webkitRequestFullscreen;
-        if (!requestFunction) {
-            return;
-        }
-        requestFunction.call(element);
-    }
-
-    /**
-     * Asks the browser to exit fullscreen mode
-     */
-    static _ExitFullscreen(): void {
-        const anyDoc = document as any;
-
-        if (document.exitFullscreen) {
-            document.exitFullscreen();
-        } else if (anyDoc.webkitCancelFullScreen) {
-            anyDoc.webkitCancelFullScreen();
         }
     }
