+++ packages/dev/core/src/Materials/Textures/renderTargetTexture.ts
@@ -19,11 +19,12 @@ import type { IRenderTargetTexture, RenderTargetWrapper } from "../../Engines/re
 
 import "../../Engines/Extensions/engine.renderTarget";
 import "../../Engines/Extensions/engine.renderTargetCube";
-import { Engine } from "../../Engines/engine";
 import { _ObserveArray } from "../../Misc/arrayTools";
 import { DumpTools } from "../../Misc/dumpTools";
 
 import type { Material } from "../material";
+import type { AbstractEngine } from "../../Engines/abstractEngine";
+import { FloorPOT, NearestPOT } from "../../Misc/tools.functions";
 
 /**
  * Options for the RenderTargetTexture constructor
@@ -197,7 +198,7 @@ export class RenderTargetTexture extends Texture implements IRenderTargetTexture
         return this._postProcesses;
     }
     private _postProcesses: PostProcess[];
-    private _resizeObserver: Nullable<Observer<Engine>>;
+    private _resizeObserver: Nullable<Observer<AbstractEngine>>;
 
     private get _prePassEnabled() {
         return !!this._prePassRenderTarget && this._prePassRenderTarget.enabled;
@@ -262,14 +263,14 @@ export class RenderTargetTexture extends Texture implements IRenderTargetTexture
     /**
      * An event triggered after the texture clear
      */
-    public onClearObservable = new Observable<Engine>();
+    public onClearObservable = new Observable<AbstractEngine>();
 
-    private _onClearObserver: Nullable<Observer<Engine>>;
+    private _onClearObserver: Nullable<Observer<AbstractEngine>>;
     /**
      * Set a clear callback in the texture.
      * This has been kept for backward compatibility and use of onClearObservable is recommended.
      */
-    public set onClear(callback: (Engine: Engine) => void) {
+    public set onClear(callback: (Engine: AbstractEngine) => void) {
         if (this._onClearObserver) {
             this.onClearObservable.remove(this._onClearObserver);
         }
@@ -1042,10 +1043,10 @@ export class RenderTargetTexture extends Texture implements IRenderTargetTexture
     private _bestReflectionRenderTargetDimension(renderDimension: number, scale: number): number {
         const minimum = 128;
         const x = renderDimension * scale;
-        const curved = Engine.NearestPOT(x + (minimum * minimum) / (minimum + x));
+        const curved = NearestPOT(x + (minimum * minimum) / (minimum + x));
 
         // Ensure we don't exceed the render dimension (while staying POT)
-        return Math.min(Engine.FloorPOT(renderDimension), curved);
+        return Math.min(FloorPOT(renderDimension), curved);
     }
 
     private _prepareRenderingManager(currentRenderList: Array<AbstractMesh>, currentRenderListLength: number, camera: Nullable<Camera>, checkLayerMask: boolean): void {
@@ -1146,7 +1147,7 @@ export class RenderTargetTexture extends Texture implements IRenderTargetTexture
         }
     }
 
-    protected _unbindFrameBuffer(engine: Engine, faceIndex: number): void {
+    protected _unbindFrameBuffer(engine: AbstractEngine, faceIndex: number): void {
         if (!this._renderTarget) {
             return;
         }
@@ -1289,7 +1290,7 @@ export class RenderTargetTexture extends Texture implements IRenderTargetTexture
         this._unbindFrameBuffer(engine, faceIndex);
 
         if (this._texture && this.isCube && faceIndex === 5) {
-            engine.generateMipMapsForCubemap(this._texture);
+            engine.generateMipMapsForCubemap(this._texture, true);
         }
 
         engine._debugPopGroup?.(1);
