+++ packages/dev/core/src/Misc/tools.functions.ts
@@ -23,3 +23,44 @@ export function IsExponentOfTwo(value: number): boolean {
 export function Mix(a: number, b: number, alpha: number): number {
     return a * (1 - alpha) + b * alpha;
 }
+
+/**
+ * Find the nearest power of two.
+ * @param x Number to start search from.
+ * @returns Next nearest power of two.
+ */
+export function NearestPOT(x: number): number {
+    const c = CeilingPOT(x);
+    const f = FloorPOT(x);
+    return c - x > x - f ? f : c;
+}
+
+/**
+ * Find the next highest power of two.
+ * @param x Number to start search from.
+ * @returns Next highest power of two.
+ */
+export function CeilingPOT(x: number): number {
+    x--;
+    x |= x >> 1;
+    x |= x >> 2;
+    x |= x >> 4;
+    x |= x >> 8;
+    x |= x >> 16;
+    x++;
+    return x;
+}
+
+/**
+ * Find the next lowest power of two.
+ * @param x Number to start search from.
+ * @returns Next lowest power of two.
+ */
+export function FloorPOT(x: number): number {
+    x = x | (x >> 1);
+    x = x | (x >> 2);
+    x = x | (x >> 4);
+    x = x | (x >> 8);
+    x = x | (x >> 16);
+    return x - (x >> 1);
+}
